// webapps/admin/generated/http-clients/backend/models/system-message-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Shula API
 * API documentation for the Shula application
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SystemMessageResponseDto
 */
export interface SystemMessageResponseDto {
    /**
     * Unique identifier for the system message
     * @type {string}
     * @memberof SystemMessageResponseDto
     */
    'id': string;
    /**
     * Unique key for the system message
     * @type {string}
     * @memberof SystemMessageResponseDto
     */
    'key': string;
    /**
     * Content of the system message with optional placeholders
     * @type {string}
     * @memberof SystemMessageResponseDto
     */
    'content': string;
    /**
     * Category of the system message
     * @type {string}
     * @memberof SystemMessageResponseDto
     */
    'category': SystemMessageResponseDtoCategoryEnum;
    /**
     * Additional metadata for the system message
     * @type {object}
     * @memberof SystemMessageResponseDto
     */
    'metadata': object;
    /**
     * Creation timestamp
     * @type {string}
     * @memberof SystemMessageResponseDto
     */
    'createdAt': string;
    /**
     * Last update timestamp
     * @type {string}
     * @memberof SystemMessageResponseDto
     */
    'updatedAt': string;
}

export const SystemMessageResponseDtoCategoryEnum = {
    Group: 'GROUP',
    Person: 'PERSON',
    Notification: 'NOTIFICATION',
    General: 'GENERAL'
} as const;

export type SystemMessageResponseDtoCategoryEnum = typeof SystemMessageResponseDtoCategoryEnum[keyof typeof SystemMessageResponseDtoCategoryEnum];